seperated IngridientInput component
